// src/actions/domain/enums.ts
/**
 * Auto generated. DO NOT edit manually.
 * Last updated on: Fri, 14 Jul 2023 17:28:18 GMT
 */

/**
 * @enum
 * @readonly
 * @description DaoWorlds Action Names
 */
export enum DaoWorldsActionName {
    Appointcust = 'appointcust',
    Balanceobsv = 'balanceobsv',
    Claimbudget = 'claimbudget',
    Claimpay = 'claimpay',
    Firecand = 'firecand',
    Firecust = 'firecust',
    Flagcandprof = 'flagcandprof',
    Newperiod = 'newperiod',
    Nominate = 'nominate',
    Nominatecane = 'nominatecane',
    Paycpu = 'paycpu',
    Regproxy = 'regproxy',
    Rejectcuspay = 'rejectcuspay',
    Removecand = 'removecand',
    Removecstvte = 'removecstvte',
    Removecuspay = 'removecuspay',
    Resigncust = 'resigncust',
    Runnewperiod = 'runnewperiod',
    Setbudget = 'setbudget',
    Setperm = 'setperm',
    Stakeobsv = 'stakeobsv',
    Stprofile = 'stprofile',
    Unregproxy = 'unregproxy',
    Unsetbudget = 'unsetbudget',
    Updatebio = 'updatebio',
    Updateconfige = 'updateconfige',
    Updatereqpay = 'updatereqpay',
    Votecust = 'votecust',
    Voteproxy = 'voteproxy',
    Weightobsv = 'weightobsv',
    Withdrawcane = 'withdrawcane',
};